# openapi.yaml
openapi: 3.0.1
info:
  title: Phone Configuration Plugin
  description: A plugin that will return a url that will help user change phone configuration(or do the action) depends on user\'s command,if the url is not empty means the plugin can do the configuration(action) , please display the url to user. If the link is empty means the plugin can not handle the command.
  version: 'v1'
servers:
  - url: https://currency-conversion-plugin2.hanlufeng.repl.co

paths:
  /convert:
    get:
      operationId: changeConfig
      summary: Depends on user\'s input to change smartphone settings
      parameters:
        - name: command
          in: query
          description: Users command which will configure the smartphone. The command must be one of these  [Mute Device,unMute Device,Block Camera,Unblock Camera,Change brightness],otherwise just tell user i can not configure per your command
          required: true
          schema:
            type: string

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/changeConfigResponse'

  /setaction:
    get:
      operationId: changeConfig
      summary: Get condition and action per  user\'s requirement
      parameters:
        - name: condition
          in: query
          description:  The condition type must be  one of [time, location], otherwise just tell user i can not configure per your command.
          required: true
          schema:
            type: string
        - name: action
          in: query
          description: action is what the user want the phone to do, for example ring device, send a notification, etc
          required: true
          schema:
            type: string

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/setConfigResponse'
components:
  schemas:
    changeConfigResponse:
      type: object
      properties:
        url:
          type: string
          description: The url that will help user do the action using the phone,if the url is empty means the plugin can not do the action per user command, if the url is not empty means the plugin can help user do the action.
        command:
          type: string
          description: User input command
        description:
          type: string
          description: descript the plugin whether can handle the command or not

    setConfigResponse:
      type: object
      properties:
        url:
          type: string
          description: The url that will help the phone do the action,if the url is empty means the plugin can not help the phone do the action per user command, if the url is not empty means the plugin can not help the phone do action.
        condition:
          type: string
          description: User input condition
        action:
          type: string
          description: User input action
        description:
          type: string
          description: descript the plugin whether can handle the command or not